+++ cove/input/views.py
@@ -76,6 +76,13 @@ def data_input(request, form_classes=default_form_classes, text_file_name='test.
             if form_name == 'url_form':
                 try:
                     data.download()
+                except requests.InvalidURL as err:
+                    return render(request, 'error.html', context={
+                        'sub_title': _("That URL is invalid"),
+                        'link': 'index',
+                        'link_text': _('Try Again'),
+                        'msg': str(err)
+                    })
                 except requests.ConnectionError as err:
                     return render(request, 'error.html', context={
                         'sub_title': _("Sorry we got a ConnectionError whilst trying to download that file"),